server/requirements.txt
scapy==2.5.0
scikit-learn==1.5.0
joblib==1.3.2
aiohttp==3.9.0
PyJWT==2.8.0
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
pydantic==2.5.0
requests==2.31.0